+++ lib/rubocop.rb
@@ -86,6 +86,7 @@ require 'rubocop/cop/rails/validation'
 require 'rubocop/formatter/base_formatter'
 require 'rubocop/formatter/plain_text_formatter'
 require 'rubocop/formatter/emacs_style_formatter'
+require 'rubocop/formatter/progress_formatter'
 
 require 'rubocop/config'
 require 'rubocop/config_store'