+++ spec/rubocop/cops/offence_spec.rb
@@ -35,6 +35,26 @@ module Rubocop
 
         expect(o1 == o2).to be_true
       end
+
+      describe '#severity_level' do
+        subject(:severity_level) do
+          Offence.new(severity, 1, 'message', 'CopName').severity_level
+        end
+
+        context 'when severity is :refactor' do
+          let(:severity) { :refactor }
+          it 'is 1' do
+            expect(severity_level).to eq(1)
+          end
+        end
+
+        context 'when severity is :fatal' do
+          let(:severity) { :fatal }
+          it 'is 5' do
+            expect(severity_level).to eq(5)
+          end
+        end
+      end
     end
   end
 end